+++ lua/lsp/handlers.lua
@@ -1,0 +1,21 @@
+-- Set Default Prefix.
+-- Note: You can set a prefix per lsp server in the lv-globals.lua file
+local M = {}
+
+function M.setup()
+  vim.lsp.handlers["textDocument/publishDiagnostics"] = vim.lsp.with(vim.lsp.diagnostic.on_publish_diagnostics, {
+    virtual_text = lvim.lsp.diagnostics.virtual_text,
+    signs = lvim.lsp.diagnostics.signs,
+    underline = lvim.lsp.document_highlight,
+  })
+
+  vim.lsp.handlers["textDocument/hover"] = vim.lsp.with(vim.lsp.handlers.hover, {
+    border = lvim.lsp.popup_border,
+  })
+
+  vim.lsp.handlers["textDocument/signatureHelp"] = vim.lsp.with(vim.lsp.handlers.signature_help, {
+    border = lvim.lsp.popup_border,
+  })
+end
+
+return M
