+++ lua/lsp/signs.lua
@@ -1,0 +1,20 @@
+local M = {}
+function M.setup()
+  vim.fn.sign_define(
+    "LspDiagnosticsSignError",
+    { texthl = "LspDiagnosticsSignError", text = "", numhl = "LspDiagnosticsSignError" }
+  )
+  vim.fn.sign_define(
+    "LspDiagnosticsSignWarning",
+    { texthl = "LspDiagnosticsSignWarning", text = "", numhl = "LspDiagnosticsSignWarning" }
+  )
+  vim.fn.sign_define(
+    "LspDiagnosticsSignHint",
+    { texthl = "LspDiagnosticsSignHint", text = "", numhl = "LspDiagnosticsSignHint" }
+  )
+  vim.fn.sign_define(
+    "LspDiagnosticsSignInformation",
+    { texthl = "LspDiagnosticsSignInformation", text = "", numhl = "LspDiagnosticsSignInformation" }
+  )
+end
+return M
